requirements.txt
PyQt6==6.6.1
pillow
ujson